Remove logic for adding default-items
Since it couldn't be 100% I removed the logic for adding default passports to the users packing list when updating location and changing country on an existing travel.
For example if an admin changed a travel it could never be correct anyway.
To be able to handle it I would need to add User to every travel.